Move files to mylib
