--- task1/Makefile
@@ -1,2 +1,0 @@
-all:
-	go run main.go book.go bookidgenerator.go bookstore.go mapstore.go slicestore.go
