+++ task1/mylib/bookidgenerator.go
@@ -1,4 +1,4 @@
-package main
+package mylib
 
 import (
 	"hash/adler32"
