+++ task1/mylib/bookstore.go
@@ -1,4 +1,4 @@
-package main
+package mylib
 
 type BookStore interface {
 	Search(id uint32) (Book, bool)
